new file:   tasca1.py 	modified:   tasca2.py 	modified:   tasca3.py 	new file:   tasca4.py 	new file:   tasca5.py 	new file:   tasca6.py 	deleted:    tasca_1.py

+++ tasca3.py
@@ -1,3 +1,4 @@
-edat = int(input('Quina edat tens?'))
+#Et demani l'edat i et diguin quants anys falten per jubilar-te Pensa que la Jubilació és als 67 anys.
+edat = float(input('Quina edat tens?'))
 jubilacio = 67
 print(f'Et falten {jubilacio- edat} anys per jubilarte.')
